update tagset for ds v3
