+++ src/components/SearchFilters/AppliedFilters.tsx
@@ -11,6 +11,7 @@ import {
 import type { Option } from "../../types/filterTypes"
 import { useContext } from "react"
 import { SearchResultsAggregationsContext } from "../../context/SearchResultsAggregationsContext"
+import { removeSelectedTag } from "./appliedFilterUtils"
 
 const AppliedFilters = () => {
   const aggregations = useContext(SearchResultsAggregationsContext)
@@ -30,36 +31,20 @@ const AppliedFilters = () => {
     })
     .flat()
 
-  const handleRemove = (tagLabel: string) => {
-    if (tagLabel === "clearFilters") {
+  const handleRemove = (tag: { label: string; field: string }) => {
+    console.log(tag)
+    if (tag.label === "Clear Filters") {
       router.push({
         pathname: "/search",
         query: getQueryWithoutFilters(router.query),
       })
       return
     }
-    // This click handler only returns the label, but we need the field so
-    // we don't remove the wrong filter
-    const relevantField = tagSetData.find(
-      (appliedFilter) => appliedFilter.label === tagLabel
-    ).field
-    const doesNotMatchLabelToRemove = (option: Option) =>
-      option.label !== tagLabel
-    // regenerate the selected options for the relevant field by removing only
-    // the tag that was selected.
-    const updatedField = appliedFiltersWithLabels[relevantField]
-      .filter(doesNotMatchLabelToRemove)
-      // only return the value so we can generate the query again
-      .map((selectedOption: Option) => selectedOption.value)
-    const updatedFilters = appliedFilters
-    // if there are elements left in the relevant field, we want those in the \
-    // query
-    if (updatedField.length) {
-      updatedFilters[relevantField] = updatedField
-      // otherwise, remove the field from the query so we don't have an
-      // undefined query param
-    } else delete updatedFilters[relevantField]
-
+    const updatedFilters = removeSelectedTag(
+      tag,
+      appliedFiltersWithLabels,
+      appliedFilters
+    )
     const updatedQuery = {
       ...getQueryWithoutFilters(router.query),
       ...buildFilterQuery(updatedFilters),
